moved categoryform.js into forms subdirectory, updated imports
